Add autolevelwidget to audio wizard
Add a new button in audio wizard, which opens a *modal* window for the
level detection widget.

The widget depends on a FFT library, which is built separately.

The widget also depends on QAudio, which in 4.x land comes from
QtMobility (more precisely: QtMultimediaKit). In 5.x the multimedia
module would be part of Qt proper.