Depend on QtMobility (YUCK)
First, add a cmake module finder for QtMobility.  File lifted from
http://qt.gitorious.org/pyside/pyside-mobility and modified to search
system paths again (file itself only looked for in-tree paths)

Then, add qtmobility-dev to Ubuntu build-deps.

Widget depends on QtMultimediaKit (which comes from mobility). This is
an evil dependency, and introduced only as a last-ditch effort to get
the widget working in the first place.

The widget *MUST* be rewritten so that it does not depend on QtMobility.
Installing that package on Linux is fairly easy, but for Windows/OSX the
new dependency is an absolute no-no.