BROKEN BUILDFIX: try to build with mobility
*WARNING*

These are atrocities against the build system, but they are a last
resort method of making the code testable - EVEN JUST ONCE.

The "fixes" are in no way complete, as the code still refuses to build
as part of tundra. But at least the errors after these changes are
different than before.