Try to grab all audio in helper
When launching the autolevel widget, it *seems* that the input audio
is enslaved to currently set clip values. Set clips suitably low so
autoleveler can do its work.

This qualifies as a weird hack, and the clip (re)sets are open-coded; in
fact, they are copied practically verbatim from other location in the
same file.

Logically this should work better. Testing in practice has still left
the jury out. But it looks promising.